refactor(core): split comments-plugin into focused modules
Break the 1.8k-line comments-plugin.ts into single-purpose files so each
module owns one concern, and unify the HTTP endpoints it served behind
one Vite plugin:

- comments.ts: marker parsing, b64url helpers, JSX-insertion planning
- edit-ops.ts: applyEdit + splice builders + asset-import planning
- revert-asset.ts: applyRevertAsset, findAssetUsages, placeholder helpers
- api-plugin.ts: single plugin registering /__comments and /__edit/*
- babel-walk.ts: gains a shared parseSource helper
- Tests split into comments/edit-ops/revert-asset/request-guard files
