+++ packages/core/src/vite/comments.test.ts
@@ -1,0 +1,72 @@
+import { describe, expect, it } from 'vitest';
+import { b64urlDecode, b64urlEncode, parseMarkers } from './comments.ts';
+
+describe('b64url encoding', () => {
+  it('round-trips arbitrary unicode strings', () => {
+    const samples = ['hello', '안녕하세요', '🎉🎊', 'a/b+c=d', JSON.stringify({ note: 'hi' })];
+    for (const s of samples) {
+      expect(b64urlDecode(b64urlEncode(s))).toBe(s);
+    }
+  });
+
+  it('produces url-safe output (no +, /, or =)', () => {
+    const encoded = b64urlEncode('subject?with/lots+of==special chars');
+    expect(encoded).not.toMatch(/[+/=]/);
+  });
+
+  it('decodes the empty string', () => {
+    expect(b64urlDecode('')).toBe('');
+  });
+});
+
+describe('parseMarkers', () => {
+  it('returns no comments when the source has no markers', () => {
+    expect(parseMarkers('const a = 1;\nexport default [];\n')).toEqual([]);
+  });
+
+  it('extracts a single marker with its line number and decoded note', () => {
+    const payload = b64urlEncode(JSON.stringify({ note: 'tighten this' }));
+    const ts = '2026-04-25T00:00:00.000Z';
+    const id = 'c-deadbeef';
+    const source = [
+      'export default [() => (',
+      '  <div>',
+      `    {/* @slide-comment id="${id}" ts="${ts}" text="${payload}" */}`,
+      '    hi',
+      '  </div>',
+      ')];',
+      '',
+    ].join('\n');
+
+    const comments = parseMarkers(source);
+    expect(comments).toEqual([{ id, line: 3, ts, note: 'tighten this', hint: undefined }]);
+  });
+
+  it('extracts a hint when the marker payload includes one', () => {
+    const payload = b64urlEncode(JSON.stringify({ note: 'fix', hint: 'h1' }));
+    const source = `{/* @slide-comment id="c-12345678" ts="2026-04-25T00:00:00.000Z" text="${payload}" */}`;
+    const [c] = parseMarkers(source);
+    expect(c.hint).toBe('h1');
+    expect(c.note).toBe('fix');
+  });
+
+  it('skips markers whose payload is malformed', () => {
+    const source =
+      '{/* @slide-comment id="c-12345678" ts="2026-04-25T00:00:00.000Z" text="not_json" */}';
+    expect(parseMarkers(source)).toEqual([]);
+  });
+
+  it('extracts multiple markers from different lines', () => {
+    const p1 = b64urlEncode(JSON.stringify({ note: 'one' }));
+    const p2 = b64urlEncode(JSON.stringify({ note: 'two' }));
+    const source = [
+      `{/* @slide-comment id="c-aaaaaaaa" ts="2026-04-25T00:00:00.000Z" text="${p1}" */}`,
+      'const x = 1;',
+      `{/* @slide-comment id="c-bbbbbbbb" ts="2026-04-25T00:00:00.000Z" text="${p2}" */}`,
+    ].join('\n');
+
+    const comments = parseMarkers(source);
+    expect(comments.map((c) => c.note)).toEqual(['one', 'two']);
+    expect(comments.map((c) => c.line)).toEqual([1, 3]);
+  });
+});
