+++ packages/core/src/vite/edit-ops.test.ts
@@ -1,176 +1,5 @@
 import { describe, expect, it } from 'vitest';
-import {
-  applyEdit,
-  applyRevertAsset,
-  b64urlDecode,
-  b64urlEncode,
-  findAssetUsages,
-  parseMarkers,
-  safeAssetIdentifier,
-} from './comments-plugin.ts';
-import { validateMutationRequest } from './request-guard.ts';
-
-describe('b64url encoding', () => {
-  it('round-trips arbitrary unicode strings', () => {
-    const samples = ['hello', '안녕하세요', '🎉🎊', 'a/b+c=d', JSON.stringify({ note: 'hi' })];
-    for (const s of samples) {
-      expect(b64urlDecode(b64urlEncode(s))).toBe(s);
-    }
-  });
-
-  it('produces url-safe output (no +, /, or =)', () => {
-    const encoded = b64urlEncode('subject?with/lots+of==special chars');
-    expect(encoded).not.toMatch(/[+/=]/);
-  });
-
-  it('decodes the empty string', () => {
-    expect(b64urlDecode('')).toBe('');
-  });
-});
-
-describe('parseMarkers', () => {
-  it('returns no comments when the source has no markers', () => {
-    expect(parseMarkers('const a = 1;\nexport default [];\n')).toEqual([]);
-  });
-
-  it('extracts a single marker with its line number and decoded note', () => {
-    const payload = b64urlEncode(JSON.stringify({ note: 'tighten this' }));
-    const ts = '2026-04-25T00:00:00.000Z';
-    const id = 'c-deadbeef';
-    const source = [
-      'export default [() => (',
-      '  <div>',
-      `    {/* @slide-comment id="${id}" ts="${ts}" text="${payload}" */}`,
-      '    hi',
-      '  </div>',
-      ')];',
-      '',
-    ].join('\n');
-
-    const comments = parseMarkers(source);
-    expect(comments).toEqual([{ id, line: 3, ts, note: 'tighten this', hint: undefined }]);
-  });
-
-  it('extracts a hint when the marker payload includes one', () => {
-    const payload = b64urlEncode(JSON.stringify({ note: 'fix', hint: 'h1' }));
-    const source = `{/* @slide-comment id="c-12345678" ts="2026-04-25T00:00:00.000Z" text="${payload}" */}`;
-    const [c] = parseMarkers(source);
-    expect(c.hint).toBe('h1');
-    expect(c.note).toBe('fix');
-  });
-
-  it('skips markers whose payload is malformed', () => {
-    const source =
-      '{/* @slide-comment id="c-12345678" ts="2026-04-25T00:00:00.000Z" text="not_json" */}';
-    expect(parseMarkers(source)).toEqual([]);
-  });
-
-  it('extracts multiple markers from different lines', () => {
-    const p1 = b64urlEncode(JSON.stringify({ note: 'one' }));
-    const p2 = b64urlEncode(JSON.stringify({ note: 'two' }));
-    const source = [
-      `{/* @slide-comment id="c-aaaaaaaa" ts="2026-04-25T00:00:00.000Z" text="${p1}" */}`,
-      'const x = 1;',
-      `{/* @slide-comment id="c-bbbbbbbb" ts="2026-04-25T00:00:00.000Z" text="${p2}" */}`,
-    ].join('\n');
-
-    const comments = parseMarkers(source);
-    expect(comments.map((c) => c.note)).toEqual(['one', 'two']);
-    expect(comments.map((c) => c.line)).toEqual([1, 3]);
-  });
-});
-
-describe('validateMutationRequest', () => {
-  function makeReq(
-    headers: Record<string, string | undefined>,
-    opts: { encrypted?: boolean } = {},
-  ) {
-    return {
-      headers,
-      socket: { encrypted: opts.encrypted ?? false },
-    } as unknown as Parameters<typeof validateMutationRequest>[0];
-  }
-
-  it('accepts same-origin JSON mutation requests', () => {
-    const req = makeReq({
-      host: 'localhost:5173',
-      origin: 'http://localhost:5173',
-      'content-type': 'application/json',
-      'sec-fetch-site': 'same-origin',
-    });
-    expect(validateMutationRequest(req, { requireJsonBody: true })).toEqual({ ok: true });
-  });
-
-  it('accepts forwarded host/proto when origin matches', () => {
-    const req = makeReq({
-      host: '127.0.0.1:5173',
-      origin: 'https://slides.example.dev',
-      'x-forwarded-host': 'slides.example.dev',
-      'x-forwarded-proto': 'https',
-      'content-type': 'application/json; charset=utf-8',
-    });
-    expect(validateMutationRequest(req, { requireJsonBody: true })).toEqual({ ok: true });
-  });
-
-  it('rejects non-JSON mutation bodies when required', () => {
-    const req = makeReq({
-      host: 'localhost:5173',
-      origin: 'http://localhost:5173',
-      'content-type': 'text/plain',
-    });
-    expect(validateMutationRequest(req, { requireJsonBody: true })).toEqual({
-      ok: false,
-      status: 415,
-      error: 'content-type must be application/json',
-    });
-  });
-
-  it('rejects cross-site browser requests', () => {
-    const req = makeReq({
-      host: 'localhost:5173',
-      origin: 'http://localhost:5173',
-      'sec-fetch-site': 'cross-site',
-    });
-    expect(validateMutationRequest(req)).toEqual({
-      ok: false,
-      status: 403,
-      error: 'cross-site request blocked',
-    });
-  });
-
-  it('rejects opaque origin values', () => {
-    const req = makeReq({
-      host: 'localhost:5173',
-      origin: 'null',
-    });
-    expect(validateMutationRequest(req)).toEqual({
-      ok: false,
-      status: 403,
-      error: 'opaque origin is not allowed',
-    });
-  });
-
-  it('rejects mismatched origin/host combinations', () => {
-    const req = makeReq({
-      host: 'localhost:5173',
-      origin: 'http://evil.example',
-    });
-    expect(validateMutationRequest(req)).toEqual({
-      ok: false,
-      status: 403,
-      error: 'origin mismatch',
-    });
-  });
-
-  it('allows trusted requests even when origin is missing', () => {
-    const req = makeReq({
-      host: 'localhost:5173',
-      'sec-fetch-site': 'same-origin',
-      'content-type': 'application/json',
-    });
-    expect(validateMutationRequest(req, { requireJsonBody: true })).toEqual({ ok: true });
-  });
-});
+import { applyEdit, safeAssetIdentifier } from './edit-ops.ts';
 
 describe('applyEdit / set-style', () => {
   // Every JSX opening tag in these synthetic sources sits at column 0;
@@ -1313,183 +1142,5 @@ describe('applyEdit / replace-placeholder-with-image', () => {
     expect(r.ok).toBe(false);
     if (r.ok) throw new Error('expected failure');
     expect(r.error).toMatch(/\.\/assets\//);
-  });
-});
-
-describe('findAssetUsages', () => {
-  it('returns 0 when the import is absent', () => {
-    const src = [
-      "import { ImagePlaceholder } from '@open-slide/core';",
-      'export default [() => (<ImagePlaceholder hint="x" />)];',
-      '',
-    ].join('\n');
-    expect(findAssetUsages(src, './assets/hero.png')).toBe(0);
-  });
-
-  it('counts every <img src={ident}> that references the import', () => {
-    const src = [
-      "import hero from './assets/hero.png';",
-      'export default [() => (',
-      '  <div>',
-      "    <img src={hero} alt='a' style={{ objectFit: 'cover' }} />",
-      "    <img src={hero} alt='b' style={{ objectFit: 'cover' }} />",
-      '  </div>',
-      ')];',
-      '',
-    ].join('\n');
-    expect(findAssetUsages(src, './assets/hero.png')).toBe(2);
-  });
-
-  it('ignores <img src> that points at a different identifier', () => {
-    const src = [
-      "import hero from './assets/hero.png';",
-      "import logo from './assets/logo.svg';",
-      'export default [() => (',
-      "  <img src={logo} alt='l' />",
-      ')];',
-      '',
-    ].join('\n');
-    expect(findAssetUsages(src, './assets/hero.png')).toBe(0);
-  });
-});
-
-describe('applyRevertAsset', () => {
-  it('reverts <img> back to <ImagePlaceholder> and removes the asset import', () => {
-    const src = [
-      "import { ImagePlaceholder } from '@open-slide/core';",
-      "import hero from './assets/hero.png';",
-      'export default [() => (',
-      "  <img src={hero} alt='Product hero' style={{ width: 1280, height: 720, objectFit: 'cover', objectPosition: '50% 50%' }} />",
-      ')];',
-      '',
-    ].join('\n');
-    const r = applyRevertAsset(src, './assets/hero.png');
-    if (!r.ok) throw new Error(`expected ok, got ${r.error}`);
-    expect(r.source).not.toContain("from './assets/hero.png'");
-    expect(r.source).not.toContain('<img');
-    expect(r.source).toContain(
-      '<ImagePlaceholder hint="Product hero" width={1280} height={720} />',
-    );
-  });
-
-  it('omits width/height when the style only carries objectFit', () => {
-    const src = [
-      "import { ImagePlaceholder } from '@open-slide/core';",
-      "import logo from './assets/logo.svg';",
-      'export default [() => (',
-      "  <img src={logo} alt='Logo' style={{ objectFit: 'cover', objectPosition: '50% 50%' }} />",
-      ')];',
-      '',
-    ].join('\n');
-    const r = applyRevertAsset(src, './assets/logo.svg');
-    if (!r.ok) throw new Error(`expected ok, got ${r.error}`);
-    expect(r.source).toContain('<ImagePlaceholder hint="Logo" />');
-    expect(r.source).not.toMatch(/width=\{/);
-    expect(r.source).not.toMatch(/height=\{/);
-  });
-
-  it("treats '100%' style fillers as unset and keeps numeric dims", () => {
-    const src = [
-      "import { ImagePlaceholder } from '@open-slide/core';",
-      "import cover from './assets/cover.png';",
-      'export default [() => (',
-      "  <img src={cover} alt='Cover' style={{ width: 800, height: '100%', objectFit: 'cover', objectPosition: '50% 50%' }} />",
-      ')];',
-      '',
-    ].join('\n');
-    const r = applyRevertAsset(src, './assets/cover.png');
-    if (!r.ok) throw new Error(`expected ok, got ${r.error}`);
-    expect(r.source).toContain('<ImagePlaceholder hint="Cover" width={800} />');
-  });
-
-  it('adds ImagePlaceholder to an existing @open-slide/core named import', () => {
-    const src = [
-      "import { type Page } from '@open-slide/core';",
-      "import hero from './assets/hero.png';",
-      'export default [() => (',
-      "  <img src={hero} alt='x' style={{ objectFit: 'cover' }} />",
-      ')];',
-      '',
-    ].join('\n');
-    const r = applyRevertAsset(src, './assets/hero.png');
-    if (!r.ok) throw new Error(`expected ok, got ${r.error}`);
-    expect(r.source).toMatch(/import \{ type Page, ImagePlaceholder \} from '@open-slide\/core';/);
-  });
-
-  it('adds a separate value import when the only @open-slide/core import is type-only', () => {
-    const src = [
-      "import type { DesignSystem, Page, SlideMeta } from '@open-slide/core';",
-      "import hero from './assets/hero.png';",
-      'export default [() => (',
-      "  <img src={hero} alt='x' style={{ objectFit: 'cover' }} />",
-      ')];',
-      '',
-    ].join('\n');
-    const r = applyRevertAsset(src, './assets/hero.png');
-    if (!r.ok) throw new Error(`expected ok, got ${r.error}`);
-    expect(r.source).toMatch(
-      /import type \{ DesignSystem, Page, SlideMeta \} from '@open-slide\/core';/,
-    );
-    expect(r.source).toContain("import { ImagePlaceholder } from '@open-slide/core';");
-  });
-
-  it('adds a fresh @open-slide/core import when none exists', () => {
-    const src = [
-      "import hero from './assets/hero.png';",
-      'export default [() => (',
-      "  <img src={hero} alt='x' style={{ objectFit: 'cover' }} />",
-      ')];',
-      '',
-    ].join('\n');
-    const r = applyRevertAsset(src, './assets/hero.png');
-    if (!r.ok) throw new Error(`expected ok, got ${r.error}`);
-    expect(r.source.split('\n')[0]).toBe("import { ImagePlaceholder } from '@open-slide/core';");
-  });
-
-  it('is a no-op when the asset is not imported', () => {
-    const src = [
-      "import { ImagePlaceholder } from '@open-slide/core';",
-      'export default [() => (<ImagePlaceholder hint="x" />)];',
-      '',
-    ].join('\n');
-    const r = applyRevertAsset(src, './assets/missing.png');
-    if (!r.ok) throw new Error(`expected ok, got ${r.error}`);
-    expect(r.source).toBe(src);
-  });
-
-  it('reverts multiple <img> usages of the same import in one file', () => {
-    const src = [
-      "import { ImagePlaceholder } from '@open-slide/core';",
-      "import hero from './assets/hero.png';",
-      'export default [() => (',
-      '  <div>',
-      "    <img src={hero} alt='First' style={{ width: 100, height: 100, objectFit: 'cover' }} />",
-      "    <img src={hero} alt='Second' style={{ objectFit: 'cover' }} />",
-      '  </div>',
-      ')];',
-      '',
-    ].join('\n');
-    const r = applyRevertAsset(src, './assets/hero.png');
-    if (!r.ok) throw new Error(`expected ok, got ${r.error}`);
-    expect(r.source).not.toContain("from './assets/hero.png'");
-    expect(r.source).toContain('<ImagePlaceholder hint="First" width={100} height={100} />');
-    expect(r.source).toContain('<ImagePlaceholder hint="Second" />');
-  });
-
-  it('refuses to revert when the identifier is referenced outside <img src>', () => {
-    const src = [
-      "import { ImagePlaceholder } from '@open-slide/core';",
-      "import hero from './assets/hero.png';",
-      'export default [() => (',
-      '  <div data-bg={hero}>',
-      "    <img src={hero} alt='x' style={{ objectFit: 'cover' }} />",
-      '  </div>',
-      ')];',
-      '',
-    ].join('\n');
-    const r = applyRevertAsset(src, './assets/hero.png');
-    expect(r.ok).toBe(false);
-    if (r.ok) throw new Error('expected failure');
-    expect(r.error).toMatch(/outside <img/);
   });
 });
